wip: default network datasource
