+++ packages/angular/src/data/http-request.builder.ts
@@ -1,37 +1,18 @@
-import { HttpClient, HttpHeaders, HttpResponse } from '@angular/common/http';
-import { BlankMapper, Mapper, ParameterType, UrlBuilder } from '@mobilejazz/harmony-core';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { ParameterType, UrlBuilder, HttpRequestBuilder } from '@mobilejazz/harmony-core';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { AuthService } from './auth.service';
 
-// TODO: Move to `harmony-core`
-interface RequestOptions {
+export interface RequestOptions {
     headers: HttpHeaders;
     observe: 'response';
     responseType: 'json';
 }
 
-// TODO: Move to `harmony-core`
-export interface HttpRequestBuilder<T = unknown> {
-    // Builder methods
-    setBody(body: unknown): this;
-    setMapper<From>(mapper: Mapper<From, T>): this;
-    setQueryParameters(queryParameters: Record<string, ParameterType>): this;
-    setUrlParameters(urlParameters: Record<string, ParameterType>): this;
-
-    // Methods
-    get(): Observable<T | undefined>;
-    post(): Observable<T | undefined>;
-    put(): Observable<T | undefined>;
-    delete(): Observable<void>;
-}
-
 export class AngularHttpRequestBuilder<T = unknown> implements HttpRequestBuilder<T> {
     private urlBuilder: UrlBuilder;
     private body: string | FormData = '';
-
-    // This `as ...` is needed to appease TS. Feel free to try to remove it.
-    private mapper: Mapper<unknown, T> = new BlankMapper() as Mapper<unknown, T>;
 
     constructor(
         protected readonly endpoint: string,
@@ -40,8 +21,6 @@ export class AngularHttpRequestBuilder<T = unknown> implements HttpRequestBuilde
     ) {
         this.urlBuilder = new UrlBuilder(endpoint);
     }
-
-    // HEADERS & OPTIONS
 
     private createRequestOptions(): RequestOptions {
         const headers = new HttpHeaders({
@@ -63,19 +42,6 @@ export class AngularHttpRequestBuilder<T = unknown> implements HttpRequestBuilde
         };
     }
 
-    public setMapper<From>(mapper: Mapper<From, T>): this {
-        this.mapper = mapper;
-        return this;
-    }
-
-    private mapResponse(res: HttpResponse<T>): T | undefined {
-        if (!res.body) {
-            return undefined;
-        }
-
-        return this.mapper.map(res.body);
-    }
-
     public setUrlParameters(urlParameters: Record<string, ParameterType>): this {
         this.urlBuilder.setUrlParameters(urlParameters);
         return this;
@@ -91,23 +57,22 @@ export class AngularHttpRequestBuilder<T = unknown> implements HttpRequestBuilde
         return this;
     }
 
-    // METHODS
     public get(): Observable<T | undefined> {
         return this.http
             .get<T>(this.urlBuilder.getUrl(), this.createRequestOptions())
-            .pipe(map((res) => this.mapResponse(res)));
+            .pipe(map((res) => res.body ?? undefined));
     }
 
     public post(): Observable<T | undefined> {
         return this.http
-            .post<T>(this.urlBuilder.getUrl(), this.body, this.createRequestOptions())
-            .pipe(map((res) => this.mapResponse(res)));
+            .put<T>(this.urlBuilder.getUrl(), this.body, this.createRequestOptions())
+            .pipe(map((res) => res.body ?? undefined));
     }
 
     public put(): Observable<T | undefined> {
         return this.http
             .put<T>(this.urlBuilder.getUrl(), this.body, this.createRequestOptions())
-            .pipe(map((res) => this.mapResponse(res)));
+            .pipe(map((res) => res.body ?? undefined));
     }
 
     public delete(): Observable<void> {
