+++ packages/angular/src/utilities.ts
@@ -1,5 +1,6 @@
 import { Provider, FactoryProvider, Type } from '@angular/core';
-import { HarmonyProvider } from '@mobilejazz/harmony-core';
+import { HarmonyProvider, DataSourceMapper, DefaultNetworkDataSource } from '@mobilejazz/harmony-core';
+import { ApiRequestService } from './data/api-request.service';
 
 export function createAngularProviders(provider: FactoryProvider, interactors: Type<unknown>[]): Provider[] {
     const providers: Provider[] = [provider];
@@ -15,4 +16,9 @@ export function createAngularProviders(provider: FactoryProvider, interactors: T
     });
 
     return providers;
+}
+
+export function getDefaultNetworkDataSourceMapper<T>(requestService: ApiRequestService): DataSourceMapper<T> {
+    const dataSource = new DefaultNetworkDataSource<T>(requestService);
+    return new DataSourceMapper<T>(dataSource);
 }
